fix: separate ad config for article, topic and author profile
